Separate utils into individual modules
