+++ cogs/send_introduction_reminders.py
@@ -11,17 +11,20 @@ from discord.ext import tasks
 from discord.ui import View
 from django.core.exceptions import ValidationError
 
-from cogs._utils import ErrorCaptureDecorators, TeXBotCog, capture_guild_does_not_exist_error
 from config import settings
 from db.core.models import (
     IntroductionReminderOptOutMember,
     SentOneOffIntroductionReminderMember,
 )
 from exceptions import GuestRoleDoesNotExist, UserNotInCSSDiscordServer
-from utils import TeXBot
+from utils import TeXBot, TeXBotBaseCog
+from utils.error_capture_decorators import (
+    ErrorCaptureDecorators,
+    capture_guild_does_not_exist_error,
+)
 
 
-class SendIntroductionRemindersTaskCog(TeXBotCog):
+class SendIntroductionRemindersTaskCog(TeXBotBaseCog):
     """Cog class that defines the send_introduction_reminders task."""
 
     def __init__(self, bot: TeXBot) -> None:
@@ -42,7 +45,7 @@ class SendIntroductionRemindersTaskCog(TeXBotCog):
         """
         self.send_introduction_reminders.cancel()
 
-    @TeXBotCog.listener()
+    @TeXBotBaseCog.listener()
     async def on_ready(self) -> None:
         """Add OptOutIntroductionRemindersView to the bot's list of permanent views."""
         self.bot.add_view(
